Integrated socket.io - additions needed. Removed faulty db connection
Added socket.io
Server listens for 'newOrder' and 'orderStateChange' and relays the messages to each of the clients.
Waiter component emits newOrder after a new order is submitted to the db, and listens for orderStateChange. 

Additions Needed:
Kitchen View - update the order status when: 'cooking', and 'ready'
Counter View - update order status to 'paid' when customer has paid
Waiter View - enable button when status is 'ready' which when pressed will update to 'served'